macos: fix macOS 12 conditional
The commit c7c6bc0112904e38bff30fb4ed81ecefc0db8fca uses
`MAC_OS_X_VERSION_MAX_ALLOWED` to determine whether to use the old or
new name. This definition can be 120000 even if compiling on macOS 11,
resulting in compiler warnings and a broken build. It can also be true
when compiling on macOS 12 but targeting macOS 11, which leads to a
crash when running it on macOS 11.

`MAC_OS_X_VERSION_MIN_REQUIRED` should be used instead, like ike in this patch:

https://bug-224632-attachments.webkit.org/attachment.cgi?id=426166 (from
https://bugs.webkit.org/show_bug.cgi?id=224632).

Note: the vendored hidapi and libusb libs are out of date. The
upstream versions simply replaced the `kIOMasterPortDefault` with
`0` (the official definition) to avoid this problem altogether.